requirements.txt
apiai==1.2.3
astroid==1.6.3
certifi==2018.4.16
chardet==3.0.4
dnspython==1.15.0
dnspython3==1.15.0
future==0.16.0
idna==2.6
isort==4.3.4
lazy-object-proxy==1.3.1
lxml==4.2.1
mccabe==0.6.1
numpy==1.14.2
pymongo==3.6.1
pyTelegramBotAPI==3.6.2
python-dateutil==2.7.2
python-telegram-bot==10.0.2
requests==2.18.4
six==1.11.0
urllib3==1.22
wrapt==1.10.11
